Update README.md
This is the tribute page for our Missile Man of The India - Dr. A.P.J. Abdul Kalam . He was also known as People's President.
This shows the journey of Small boy to the President of the India and also one of the most loved man in India. We should follow his steps to become successful in life and do something for our nation.
There are many quotes given by Sir Kalam that are life-changing. I like to mention one of them :-
"It Is Very Easy To Defeat Someone, But It Is Very Hard To Win Someone"